requirements.txt
tensorflow
scikit-learn
matplotlib
pillow
streamlit
joblib
numpy